first draft to be able to scale in to 0 agent
